+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/CameraSensorActivity.java
@@ -320,8 +320,10 @@ public class CameraSensorActivity extends AppCompatActivity
         ((TextView) findViewById(R.id.testTitleTextView)).setText(mainApp.currentTestInfo.getName(conf.locale.getLanguage()));
         ((TextView) findViewById(R.id.testTypeTextView)).setText(mainApp.currentTestInfo.getName(conf.locale.getLanguage()));
 
-        if (mainApp.currentTestInfo.getCode().isEmpty()) {
-            AlertUtils.showError(this, R.string.error, getString(R.string.errorLoadingTestTypes), null, R.string.ok,
+        if (!mainApp.currentTestInfo.getCode().isEmpty()) {
+
+            String message = String.format("%s\r\n\r\n%s", getString(R.string.errorLoadingConfiguration), getString(R.string.pleaseContactSupport));
+            AlertUtils.showError(this, R.string.error, message, null, R.string.ok,
                     new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialogInterface, int i) {